rodowizardv/api/management/commands/inittokens.py
from django.core.management.base import BaseCommand
from django.conf import settings
from django.contrib.auth.models import User
from rest_framework.authtoken.models import Token


class Command(BaseCommand):
    def handle(self, *args, **kwargs):
        for user in User.objects.all():
            Token.objects.get_or_create(user=user)
